Refactor import paths for TextCopyBlock component in Alerts and Description files
